Support API DsgQueryDesensStatusList.

+++ dataworks-public-20200518/core/Models/DsgQueryDefaultTemplatesRequest.cs
@@ -18,7 +18,7 @@ namespace AlibabaCloud.SDK.Dataworks_public20200518.Models
         /// </summary>
         [NameInMap("SceneId")]
         [Validation(Required=false)]
-        public int? SceneId { get; set; }
+        public long? SceneId { get; set; }
 
     }
